requirements.txt
fastapi
uvicorn
jieba
mecab-python3
pymongo
python-dotenv